+++ src/state/self.ts
@@ -1,5 +1,6 @@
 import { reactive } from 'vue';
 
 export const self = reactive({
-  slot: ''
+  slot: '',
+  url: ''
 });