+++ src/views/Login.vue
@@ -106,6 +106,7 @@ async function connect() {
   AppStorage.set('password', password.value);
 
   self.slot = slot.value;
+  self.url = url.value;
 
   loadLocations();
 
@@ -123,8 +124,7 @@ async function cancel() {
   window.location.reload();
 }
 
-// const hasMultipleAccounts = computed(() => localAccounts.value.length > 1);
-const hasMultipleAccounts = computed(() => true);
+const hasMultipleAccounts = computed(() => localAccounts.value.length > 1);
 const switchAccountList = computed(() => localAccounts.value);
 
 const accountSwitcher = ref(-1);