+++ vite.config.ts
@@ -1,6 +1,7 @@
 import { fileURLToPath, URL } from 'node:url';
 import { defineConfig } from 'vite';
 import vue from '@vitejs/plugin-vue';
+import svgLoader from 'vite-svg-loader'
 import monacoEditorPlugin from 'vite-plugin-monaco-editor';
 
 const _monacoEditorPlugin = monacoEditorPlugin as any;
@@ -10,6 +11,7 @@ export default defineConfig({
   base: './',
   plugins: [
     vue(),
+    svgLoader(),
     _monacoEditorPlugin.default({
       languages: ['css']
     })
@@ -19,4 +21,7 @@ export default defineConfig({
       '@': fileURLToPath(new URL('./src', import.meta.url))
     },
   },
-})
+  server: {
+    allowedHosts: true
+  }
+});